feat: enable pipeline and uploading application
