+++ pom.xml
@@ -62,6 +62,11 @@
       <groupId>org.jenkins-ci.plugins</groupId>
       <artifactId>credentials</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.jenkins-ci.plugins.workflow</groupId>
+      <artifactId>workflow-step-api</artifactId>
+      <scope>compile</scope>
+    </dependency>
   </dependencies>
 
   <repositories>
