+++ src/main/resources/io/jenkins/plugins/dogu/builder/DoguApplicationBuilder/config.jelly
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+    <f:entry title="Project ID" field="projectId">
+        <f:textbox/>
+    </f:entry>
+    <f:entry title="Application Path" field="applicationPath">
+        <f:textbox/>
+    </f:entry>
+    <f:entry title="Credentials" field="credentialsId">
+        <f:select/>
+    </f:entry>
+</j:jelly>
